Change the datatype for Celsius var from int to double to have a proper conversion from Farhenheit

+++ 9465_Stefano_tut2/src/StefanoTut2/Tutorial2Task1.java
@@ -9,7 +9,7 @@ public class Tutorial2Task1
     {
         String fahrenheit = JOptionPane.showInputDialog(new JFrame(),"Enter the temperature in Fahrenheit: ");
         
-        int celsius = (((Integer.valueOf(fahrenheit)-32) *5) / 9);
+        double celsius = (((Integer.valueOf(fahrenheit)-32) *5) / 9);
         
         JOptionPane.showMessageDialog(new JFrame(),"the Fahrenheit temperature in Celcius is: " + celsius);
     }
